refactor: refactors changes Graph component to new code style
